feat: TRM-34196: Move solr client related mocking to MockSolrClientConfig so that they can be used on demand
