+++ common-rest/src/test/java/org/uniprot/api/rest/CommonDataStoreTestConfig.java
@@ -1,11 +1,5 @@
 package org.uniprot.api.rest;
 
-import static org.mockito.Mockito.mock;
-
-import java.net.URISyntaxException;
-
-import org.apache.http.client.HttpClient;
-import org.apache.solr.client.solrj.SolrClient;
 import org.apache.solr.client.solrj.request.json.JsonQueryRequest;
 import org.apache.solr.common.params.ModifiableSolrParams;
 import org.springframework.boot.test.context.TestConfiguration;
@@ -15,26 +9,10 @@ import org.springframework.web.client.RestTemplate;
 import org.uniprot.api.common.repository.search.SolrRequest;
 import org.uniprot.api.common.repository.search.SolrRequestConverter;
 
+import static org.mockito.Mockito.mock;
+
 @TestConfiguration
 public class CommonDataStoreTestConfig {
-
-    @Bean
-    @Profile("offline")
-    public HttpClient httpClient() {
-        return mock(HttpClient.class);
-    }
-
-    @Bean
-    @Profile("offline")
-    public SolrClient solrClient() throws URISyntaxException {
-        return mock(SolrClient.class);
-    }
-
-    @Bean("uniProtKBSolrClient")
-    @Profile("offline")
-    public SolrClient uniProtKBSolrClient() throws URISyntaxException {
-        return mock(SolrClient.class);
-    }
 
     @Bean
     @Profile("offline")
